Use `yield...` instead of `with...yield` for building HTML
`with...yield` is known to have a few [issues][#1]. It also
broke encapsulation, in that private methods of a markout
object were accessible from within the block passed to it.

This new way avoids possible namespace collisions, since methods
are called with receiver.

I'm still open to suggestions around this.

[#1]: https://github.com/crystal-lang/crystal/issues/4436